requirements.txt
bcrypt==4.1.2
beautifulsoup4==4.12.0
cryptography==42.0.5
Flask==3.0.2
flask-cors==4.0.0
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
kafka-python==2.0.2
langchain==0.1.10
langchain-google-genai==0.0.11
langchain-openai==0.0.8
langchain-pinecone==0.0.3
Markdown==3.5.2
pillow==10.2.0
pinecone-client==3.1.0
playwright==1.42.0
pymongo==4.6.2
pyjwt==2.8.0
python-dotenv==1.0.1
qstash-python==1.0.0
requests==2.31.0
selenium==4.18.1
unstructured==0.12.5
webdriver-manager==4.0.1
werkzeug==3.0.1